+++ Cargo.toml
@@ -4,6 +4,8 @@ members = [
   "crates/settings",
   "crates/networks",
   "crates/ws",
+  "crates/peers",
+  "crates/rpc",
   "crates/abis",
   "crates/core",
   "crates/types",
@@ -29,6 +31,7 @@ iron-settings = { path = "crates/settings" }
 iron-networks = { path = "crates/networks" }
 iron-ws = { path = "crates/ws" }
 iron-peers = { path = "crates/peers" }
+iron-rpc = { path = "crates/rpc" }
 iron-types = { path = "crates/types" }
 iron-abis = { path = "crates/abis" }
 iron-db = { path = "crates/db" }
