+++ crates/peers/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "iron-networks"
+name = "iron-peers"
 version.workspace = true
 edition.workspace = true
 license-file.workspace = true
@@ -10,10 +10,6 @@ authors.workspace = true
 
 [dependencies]
 iron-types = { workspace = true }
-iron-peers = { workspace = true }
-iron-db = { workspace = true }
-iron-sync-anvil = { workspace = true }
-ethers = { workspace = true }
 tauri = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
