+++ crates/peers/src/global.rs
