+++ crates/rpc/src/error.rs
