+++ crates/rpc/src/sign_message.rs
